requirements.txt
flask
werkzeug
pdfminer.six
nltk
numpy
requests
spacy
six